requirements.txt
facebook-sdk
tweepy
schedule
beautifulsoup4
markovify
Pillow
discord-webhook